fixed: edit and remove buttons

+++ src/main/java/shop/dao/jdbc/mapper/FoodMapper.java
@@ -1,14 +1,13 @@
 package shop.dao.jdbc.mapper;
 
 import shop.web.entity.Food;
-
 import java.sql.ResultSet;
 import java.sql.SQLException;
 import java.sql.Timestamp;
 
 public class FoodMapper {
     public Food mapRow(ResultSet resultSet) throws SQLException {
-        int id = resultSet.getInt("id");
+        long id = resultSet.getLong("id");
         String name = resultSet.getString("name");
         String comment = resultSet.getString("comment");
         int price = resultSet.getInt("price");
